+++ src/sumui/__init__.py
@@ -10,20 +10,21 @@
 #  (at your option) any later version.
 #  
 
-__version__ = "0.1.0a7";
+__version__ = "0.1.0a8";
 
 from .backend import BackendCapabilities;
 from .application import BACKEND_NAMES, ActionSpec, FunctionKeySpec, MenuEntrySpec, MenuSpec, add_backend_arguments, backend_from_args, normalize_backend_name;
 from .charts import AxisSpec, ChartSeries, ChartSpec, coerce_chart_spec;
 from .events import EVENT_TYPES, UIEvent;
 from .dialogio import load_dialog_spec, parse_dialog_spec;
-from .graphics import ColorSpec, GraphicsCommand, GraphicsMode, GraphicsProgram, ImageSpec, TableSpec, basic_mode, modern_mode, spectrum_mode;
+from .graphics import ColorSpec, GraphicsCommand, GraphicsMode, GraphicsProgram, ImageSpec, TableSpec, basic_mode, display_mode, modern_mode, screen_mode, spectrum_mode;
 from .specs import DialogSpec, FieldSpec, InputSpec, MenuItemSpec, normalize_allowed_values;
 from .datetime_widgets import CalendarModel, DateTimeModel, TimeModel;
+from .typography import FontSpec;
 
 __all__ = [
     "__version__", "BackendCapabilities", "BACKEND_NAMES", "ActionSpec", "FunctionKeySpec", "MenuEntrySpec", "MenuSpec", "add_backend_arguments", "backend_from_args", "normalize_backend_name", "AxisSpec", "ChartSeries", "ChartSpec", "coerce_chart_spec",
     "EVENT_TYPES", "UIEvent", "load_dialog_spec", "parse_dialog_spec", "ColorSpec", "GraphicsCommand", "GraphicsMode", "GraphicsProgram", "ImageSpec", "TableSpec",
-    "basic_mode", "modern_mode", "spectrum_mode", "DialogSpec", "FieldSpec", "InputSpec", "MenuItemSpec",
-    "normalize_allowed_values", "CalendarModel", "TimeModel", "DateTimeModel",
+    "basic_mode", "display_mode", "modern_mode", "screen_mode", "spectrum_mode", "DialogSpec", "FieldSpec", "InputSpec", "MenuItemSpec",
+    "normalize_allowed_values", "CalendarModel", "TimeModel", "DateTimeModel", "FontSpec",
 ];